Implemented delete from s3 method
After the file is uploaded and processed for summarization, it is permanently stored in S3. This delete method ensures that the document is temporarily stored for processing and then removed from S3 when it is not needed. This may be a temporary solution for now before attempting to generate unique strings for the file name rather than pass the file name itself.